fix: allow value of a var to be push to an array
The display on the calculator indicates the correct value, however, the
value of the second variable will concatenate on the first variable.

So added a fix that will push the current value of a variable to an array.
This prevents the value from being changed with every number button click.