fix: enable user to continuously evaluate numbers
Using a second operator the user inputs a pair of numbers will evaluate
the expression. The problem was continuously evaluating numbers this way
will cause the number to concatenate to the next input.

number button listeners were moved to the global scope. Additionally,
the listener for the operator symbols has its condition in reverse. That
after evaluating a true condition, the next condition statement will
also became true.That is why the statements were put in reverse so that
evaluating a condition will not make the other statements inside the
function to became true.