fix: allow changing of the desired user operator
When the user inputs a number and repeatedly press an operator will
cause the calculator to not work. Changes were made so that any change
in user input will not overpopulate the array through which the
expression is being evaluated.

A good side effect of this is that the array will now be fixed at an
arr.length of 3 as values will not be push to the array haphazardly.