fix: reinitialize values after operate
Previously, when the user presses the '=' button, pressing buttons other
than the operators will cause the number inputted to concatenate to the
result.

To reinitialize the values, a boolean is written to check if the operate
button was registered as an input. If so, pressing buttons other than
the operator will cause all values to reinitialize.